[BE] refactor : 미팅, 참여 정보 Application 계층 의존성 줄이기
