+++ backend/src/main/java/moim_today/application/meeting/joined_meeting/JoinedMeetingServiceImpl.java
@@ -2,8 +2,8 @@ package moim_today.application.meeting.joined_meeting;
 
 import moim_today.global.error.BadRequestException;
 import moim_today.implement.meeting.joined_meeting.JoinedMeetingComposition;
-import moim_today.implement.meeting.meeting.MeetingFinder;
-import moim_today.implement.moim.moim.MoimFinder;
+import moim_today.implement.meeting.meeting.MeetingComposition;
+import moim_today.implement.moim.moim.MoimComposition;
 import moim_today.implement.schedule.schedule.ScheduleComposition;
 import moim_today.persistence.entity.meeting.meeting.MeetingJpaEntity;
 import moim_today.persistence.entity.schedule.schedule.ScheduleJpaEntity;
@@ -19,17 +19,17 @@ import static moim_today.global.constant.exception.ScheduleExceptionConstant.*;
 public class JoinedMeetingServiceImpl implements JoinedMeetingService {
 
     private final JoinedMeetingComposition joinedMeetingComposition;
-    private final MoimFinder moimFinder;
-    private final MeetingFinder meetingFinder;
+    private final MoimComposition moimComposition;
+    private final MeetingComposition meetingComposition;
     private final ScheduleComposition scheduleComposition;
 
     public JoinedMeetingServiceImpl(final JoinedMeetingComposition joinedMeetingComposition,
-                                    final MoimFinder moimFinder,
-                                    final MeetingFinder meetingFinder,
+                                    final MoimComposition moimComposition,
+                                    final MeetingComposition meetingComposition,
                                     final ScheduleComposition scheduleComposition) {
         this.joinedMeetingComposition = joinedMeetingComposition;
-        this.moimFinder = moimFinder;
-        this.meetingFinder = meetingFinder;
+        this.moimComposition = moimComposition;
+        this.meetingComposition = meetingComposition;
         this.scheduleComposition = scheduleComposition;
     }
 
@@ -41,9 +41,9 @@ public class JoinedMeetingServiceImpl implements JoinedMeetingService {
     @Transactional
     @Override
     public void acceptanceJoinMeeting(final long memberId, final long meetingId, final LocalDateTime currentDateTime) {
-        MeetingJpaEntity meetingJpaEntity = meetingFinder.getById(meetingId);
+        MeetingJpaEntity meetingJpaEntity = meetingComposition.getById(meetingId);
         meetingJpaEntity.validateCurrentTime(currentDateTime);
-        String moimTitle = moimFinder.getTitleById(meetingJpaEntity.getMoimId());
+        String moimTitle = moimComposition.getTitleById(meetingJpaEntity.getMoimId());
         ScheduleJpaEntity scheduleJpaEntity = ScheduleJpaEntity.toEntity(memberId, moimTitle, meetingJpaEntity);
         boolean isNew = scheduleComposition.createScheduleIfNotExist(scheduleJpaEntity);
 
@@ -58,7 +58,7 @@ public class JoinedMeetingServiceImpl implements JoinedMeetingService {
     @Transactional
     @Override
     public void refuseJoinMeeting(final long memberId, final long meetingId, final LocalDateTime currentDateTime) {
-        MeetingJpaEntity meetingJpaEntity = meetingFinder.getById(meetingId);
+        MeetingJpaEntity meetingJpaEntity = meetingComposition.getById(meetingId);
         meetingJpaEntity.validateCurrentTime(currentDateTime);
         boolean attendance = false;
         joinedMeetingComposition.updateAttendance(memberId, meetingId, attendance);
