+++ backend/src/main/java/moim_today/application/meeting/meeting_comment/MeetingCommentServiceImpl.java
@@ -4,10 +4,7 @@ import moim_today.dto.meeting.meeting_comment.MeetingCommentCreateRequest;
 import moim_today.dto.meeting.meeting_comment.MeetingCommentResponse;
 import moim_today.dto.meeting.meeting_comment.MeetingCommentUpdateRequest;
 import moim_today.implement.meeting.meeting.MeetingComposition;
-import moim_today.implement.meeting.meeting_comment.MeetingCommentAppender;
-import moim_today.implement.meeting.meeting_comment.MeetingCommentFinder;
-import moim_today.implement.meeting.meeting_comment.MeetingCommentRemover;
-import moim_today.implement.meeting.meeting_comment.MeetingCommentUpdater;
+import moim_today.implement.meeting.meeting_comment.*;
 import moim_today.implement.moim.joined_moim.JoinedMoimComposition;
 import moim_today.persistence.entity.meeting.meeting_comment.MeetingCommentJpaEntity;
 import org.springframework.stereotype.Service;
@@ -18,23 +15,14 @@ import java.util.List;
 @Service
 public class MeetingCommentServiceImpl implements MeetingCommentService {
 
-    private final MeetingCommentAppender meetingCommentAppender;
-    private final MeetingCommentFinder meetingCommentFinder;
-    private final MeetingCommentUpdater meetingCommentUpdater;
-    private final MeetingCommentRemover meetingCommentRemover;
+    private final MeetingCommentComposition meetingCommentComposition;
     private final MeetingComposition meetingComposition;
     private final JoinedMoimComposition joinedMoimComposition;
 
-    public MeetingCommentServiceImpl(final MeetingCommentAppender meetingCommentAppender,
-                                     final MeetingCommentFinder meetingCommentFinder,
-                                     final MeetingCommentUpdater meetingCommentUpdater,
-                                     final MeetingCommentRemover meetingCommentRemover,
+    public MeetingCommentServiceImpl(final MeetingCommentComposition meetingCommentComposition,
                                      final MeetingComposition meetingComposition,
                                      final JoinedMoimComposition joinedMoimComposition) {
-        this.meetingCommentAppender = meetingCommentAppender;
-        this.meetingCommentFinder = meetingCommentFinder;
-        this.meetingCommentUpdater = meetingCommentUpdater;
-        this.meetingCommentRemover = meetingCommentRemover;
+        this.meetingCommentComposition = meetingCommentComposition;
         this.meetingComposition = meetingComposition;
         this.joinedMoimComposition = joinedMoimComposition;
     }
@@ -43,29 +31,29 @@ public class MeetingCommentServiceImpl implements MeetingCommentService {
     public void createMeetingComment(final long memberId, final MeetingCommentCreateRequest meetingCommentCreateRequest) {
         long moimId = meetingComposition.getMoimIdByMeetingId(meetingCommentCreateRequest.meetingId());
         joinedMoimComposition.validateMemberInMoim(memberId, moimId);
-        meetingCommentAppender.createMeetingComment(memberId, meetingCommentCreateRequest);
+        meetingCommentComposition.createMeetingComment(memberId, meetingCommentCreateRequest);
     }
 
     @Override
     public List<MeetingCommentResponse> findAllByMeetingId(final long memberId, final long meetingId) {
         long moimId = meetingComposition.getMoimIdByMeetingId(meetingId);
         joinedMoimComposition.validateMemberInMoim(memberId, moimId);
-        return meetingCommentFinder.findAllByMeetingId(meetingId);
+        return meetingCommentComposition.findAllByMeetingId(meetingId);
     }
 
     @Transactional
     @Override
     public void updateMeetingComment(final long memberId, final MeetingCommentUpdateRequest meetingCommentUpdateRequest) {
         long meetingCommentId = meetingCommentUpdateRequest.meetingCommentId();
-        MeetingCommentJpaEntity meetingCommentJpaEntity = meetingCommentFinder.getById(meetingCommentId);
+        MeetingCommentJpaEntity meetingCommentJpaEntity = meetingCommentComposition.getById(meetingCommentId);
         meetingCommentJpaEntity.validateMember(memberId);
-        meetingCommentUpdater.updateMeetingComment(meetingCommentId, meetingCommentUpdateRequest);
+        meetingCommentComposition.updateMeetingComment(meetingCommentId, meetingCommentUpdateRequest);
     }
 
     @Override
     public void deleteMeetingComment(final long memberId, final long meetingCommentId) {
-        MeetingCommentJpaEntity meetingCommentJpaEntity = meetingCommentFinder.getById(meetingCommentId);
+        MeetingCommentJpaEntity meetingCommentJpaEntity = meetingCommentComposition.getById(meetingCommentId);
         meetingCommentJpaEntity.validateMember(memberId);
-        meetingCommentRemover.deleteById(meetingCommentId);
+        meetingCommentComposition.deleteById(meetingCommentId);
     }
 }
